# contracts/openapi/endpoint-platform-v1.yaml
openapi: "3.1.0"
info:
  title: "Endpoint Platform API"
  version: "v1"
paths:
...
  "/api/v1/devices/{device_id}":
    get:
      tags:
        - "endpoint-operations"
      summary: "Read Endpoint Device"
      description: "Read one exact safe device summary for verified consumer mapping."
      operationId: "read_endpoint_device_api_v1_devices__device_id__get"
      security:
        -
          ServiceBearer: []
      parameters:
        -
          name: "device_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            title: "Device Id"
        -
          name: "X-Correlation-ID"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            title: "X-Correlation-Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/DeviceSummaryEnvelope"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "404":
          description: "Device not found"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "401":
          description: "Service authentication failed"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "403":
          description: "Service scope is insufficient"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "422":
          description: "Validation Error"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      x-required-scopes:
        - "devices.read"
  "/api/v1/devices/{device_id}/capabilities":
    get:
      tags:
        - "endpoint-operations"
      summary: "Read Device Operation Capabilities"
      description: "Report only server acceptance for an active device and fixed capability."
      operationId: "read_device_operation_capabilities_api_v1_devices__device_id__capabilities_get"
      security:
        -
          ServiceBearer: []
      parameters:
        -
          name: "device_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            title: "Device Id"
        -
          name: "X-Correlation-ID"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            title: "X-Correlation-Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/DeviceCapabilitiesEnvelope"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "404":
          description: "Active device not found"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "401":
          description: "Service authentication failed"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "403":
          description: "Service scope is insufficient"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "422":
          description: "Validation Error"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      x-required-scopes:
        - "devices.read"
  "/api/v1/devices/{device_id}/operations":
    post:
      tags:
        - "endpoint-operations"
      summary: "Create Device Operation"
      description: "Create or replay one operation owned by the authenticated service client."
      operationId: "create_device_operation_api_v1_devices__device_id__operations_post"
      security:
        -
          ServiceBearer: []
      parameters:
        -
          name: "device_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            title: "Device Id"
        -
          name: "Idempotency-Key"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 8
            maxLength: 128
            pattern: "^[!-~][ -~]{6,126}[!-~]$"
            title: "Idempotency-Key"
        -
          name: "X-Correlation-ID"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            title: "X-Correlation-Id"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/EndpointOperationCreateV1"
      responses:
        "201":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OperationResponseEnvelope"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "200":
          description: "Replayed endpoint operation"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OperationResponseEnvelope"
        "409":
          description: "Idempotency key conflict"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "503":
          description: "Safe result unavailable"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "401":
          description: "Service authentication failed"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "403":
          description: "Service scope is insufficient"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "422":
          description: "Validation Error"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      x-required-scopes:
        - "operations.create"
  "/api/v1/operations/{operation_id}":
    get:
      tags:
        - "endpoint-operations"
      summary: "Read Endpoint Operation"
      description: "Read one operation through stable service-client ownership."
      operationId: "read_endpoint_operation_api_v1_operations__operation_id__get"
      security:
        -
          ServiceBearer: []
      parameters:
        -
          name: "operation_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            title: "Operation Id"
        -
          name: "X-Correlation-ID"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            title: "X-Correlation-Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OperationResponseEnvelope"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "404":
          description: "Endpoint operation not found"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "503":
          description: "Safe result unavailable"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "401":
          description: "Service authentication failed"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "403":
          description: "Service scope is insufficient"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "422":
          description: "Validation Error"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      x-required-scopes:
        - "operations.read"
  "/api/v1/operations/{operation_id}/cancel":
    post:
      tags:
        - "endpoint-operations"
      summary: "Cancel Endpoint Operation"
      description: "Cancel one owner-scoped operation only while it remains undelivered."
      operationId: "cancel_endpoint_operation_api_v1_operations__operation_id__cancel_post"
      security:
        -
          ServiceBearer: []
      parameters:
        -
          name: "operation_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            title: "Operation Id"
        -
          name: "X-Correlation-ID"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            title: "X-Correlation-Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OperationResponseEnvelope"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "404":
          description: "Endpoint operation not found"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "409":
          description: "Operation has already reached delivery or a terminal state"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "401":
          description: "Service authentication failed"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "403":
          description: "Service scope is insufficient"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "422":
          description: "Validation Error"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      x-required-scopes:
        - "operations.cancel"
...
components:
  schemas:
...
    EndpointOperationCreateV1:
      additionalProperties: false
      description: "Strict service request for one bounded diagnostic collection."
      properties:
        schema_version:
          const: "endpoint_operation_create_v1"
          title: "Schema Version"
          type: "string"
        capability:
          const: "context.diagnostic.collect"
          title: "Capability"
          type: "string"
        parameters:
          "$ref": "#/components/schemas/EndpointOperationCreateV1__DiagnosticCollectionParametersV1"
      required:
        - "schema_version"
        - "capability"
        - "parameters"
      title: "EndpointOperationCreateV1"
      type: "object"
    EndpointOperationCreateV1__DiagnosticCollectionParametersV1:
      additionalProperties: false
      description: "The only public operation input supported in v1."
      properties:
        reason:
          maxLength: 256
          minLength: 1
          pattern: "^(?![\\s\\S]*://)[^\\x00-\\x1f\\x7f]{1,256}(?![\\s\\S])"
          title: "Reason"
          type: "string"
      required:
        - "reason"
      title: "DiagnosticCollectionParametersV1"
      type: "object"
    EndpointDeviceSummaryV1:
      additionalProperties: false
      description: "Strict device projection used for external verified-reference reads."
      properties:
        schema_version:
          const: "endpoint_device_summary_v1"
          title: "Schema Version"
          type: "string"
        device_id:
          format: "uuid"
          title: "Device Id"
          type: "string"
        display_name:
          maxLength: 256
          minLength: 1
          title: "Display Name"
          type: "string"
        retired:
          title: "Retired"
          type: "boolean"
        last_seen_at:
          anyOf:
            -
              format: "date-time"
              type: "string"
              "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
            -
              type: "null"
          default: null
          title: "Last Seen At"
      required:
        - "schema_version"
        - "device_id"
        - "display_name"
        - "retired"
      title: "EndpointDeviceSummaryV1"
      type: "object"
...
    DeviceCapabilitiesEnvelope:
      properties:
        data:
          "$ref": "#/components/schemas/EndpointDeviceCapabilitiesV1"
      additionalProperties: false
      type: "object"
      required:
        - "data"
      title: "DeviceCapabilitiesEnvelope"
...
    DeviceSummaryEnvelope:
      properties:
        data:
          "$ref": "#/components/schemas/EndpointDeviceSummaryV1"
      additionalProperties: false
      type: "object"
      required:
        - "data"
      title: "DeviceSummaryEnvelope"
...
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: "array"
          title: "Detail"
      type: "object"
      title: "HTTPValidationError"
...
    OperationResponseEnvelope:
      properties:
        data:
          "$ref": "#/components/schemas/OperationResponseData"
      additionalProperties: false
      type: "object"
      required:
        - "data"
      title: "OperationResponseEnvelope"
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              -
                type: "string"
              -
                type: "integer"
          type: "array"
          title: "Location"
        msg:
          type: "string"
          title: "Message"
        type:
          type: "string"
          title: "Error Type"
      type: "object"
      required:
        - "loc"
        - "msg"
        - "type"
      title: "ValidationError"
  securitySchemes:
    AgentBearer:
      type: "http"
      scheme: "bearer"
    ServiceBearer:
      type: "http"
      scheme: "bearer"